backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
matplotlib
seaborn
shap
lime
joblib
python-multipart
python-dotenv
aiofiles
jinja2
bacon
joblib
numpy
pydantic